requirements.txt
Django~=3.2.20
djangorestframework~=3.14.0
drf-yasg~=1.21.6
coreapi~=2.3.3
psycopg2-binary~=2.9.6
django-filter~=23.2